x86_64: Move the VMCS validation code into its own module
